+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Security - in case of vulnerabilities.
 - Resource redirect after delete
 - Search indexes permissions
 - Update datasets version
+- Subsection not deleting
 
 ## [0.6.4] - 2018-10-19
 ### Added